requirements.txt
Django==2.0.2
djangorestframework==3.7.7
markdown==2.6.11
django-filter==2.0.0.dev1
beautifulsoup4==4.6.0
requests==2.18.4